+++ src/check_jsonschema/loaders/instance.py
@@ -1,17 +1,25 @@
 import json
+import typing as t
 
 import ruamel.yaml
 from identify import identify
 
+from ..transforms import TransformT
 from .errors import BadFileTypeError
 
 yaml = ruamel.yaml.YAML(typ="safe")
 
 
 class InstanceLoader:
-    def __init__(self, filenames, default_filetype=None):
+    def __init__(
+        self,
+        filenames,
+        default_filetype=None,
+        data_transform: t.Optional[TransformT] = None,
+    ):
         self._filenames = filenames
         self._default_ft = default_filetype
+        self._data_transform = data_transform
 
     @property
     def _default_loadfunc(self):
@@ -38,8 +46,15 @@ class InstanceLoader:
                 )
 
             with open(fn) as fp:
-                yield (fn, loadfunc(fp))
+                data = loadfunc(fp)
+                if self._data_transform:
+                    data = self._data_transform(data)
+                yield (fn, data)
 
 
 def instance_loader_from_args(args):
-    return InstanceLoader(args.instancefiles, default_filetype=args.default_filetype)
+    return InstanceLoader(
+        args.instancefiles,
+        default_filetype=args.default_filetype,
+        data_transform=args.data_transform,
+    )