+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 <!-- vendor-insert-here -->
+- Add support for `--data-transform azure-pipelines` to handle compile-time
+    expressions in Pipelines files. This option is applied to the azure
+    pipelines hook ([#29](https://github.com/sirosen/check-jsonschema/issues/29))
 - Use vendored schemas in all hooks, not latest schemastore copies. This
   ensures that hook behavior is consistent
   ([#38](https://github.com/sirosen/check-jsonschema/issues/38))